requirements.txt
openai
langchain
langchain-community
langchain-openai
pypdf
azure-search-documents
azure-ai-formrecognizer
python-dotenv